pyproject.toml
[project]
name = "django-oapif-demo"
version = "0.1.0"
description = "django-oapif demo project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django[argon2]==5.2",
    "django-oapif @ git+https://github.com/opengisch/django-oapif@main",
    "psycopg[binary]>=3.3.3",
    "django-unfold>=0.81.0",
    "django-leaflet>=0.33.0",
]
